Add a switch that disables redirecting standard streams to syslog. (#2684)
When developing the engine from Xcode, it is useful to see log output in Xcode's UI. We were redirecting streams away so that the tools could pick read these logs. With the change to the scheme, when running from Xcode, the logs will show up in Xcode's UI. When, running from the tools, that flag will not be present and the tools will work as they always have.