Fix resize crash in Android virtual display (#37329)
In the Virtual Display codepath for Android platform views, resize
completes asynchronously. Currently it is attempting to access the
Context in the completion handler, but there is no guarantee that it
is still present at that point, leading to possible null pointer
crashes.

This adds a check for the current state of the Context, and uses a
fallback if it's not available.

Fixes https://github.com/flutter/flutter/issues/114095