[web] - Fix for Safari partially autofilling form  (#43058)
We have an issue where autofill in Safari Desktop is only partially autofilling the form.  The only input that gets filled is the currently focused one. 

This is happening for a couple of reasons:
- We hide non-focused inputs that are part of the autofill form by sizing them to 0px height and width, among other styles.
- This doesn't cause issues in most browsers.  However, Safari does not respect offscreen or 0-sized inputs, and this leads to the partial form autofill issue. 

The solution is to make sure we hide the inputs in the rest of the form from the users, without hiding it from Safari. 

Fixes https://github.com/flutter/flutter/issues/71275