[Impeller] OES extension does not apply to regular textures for decal support (#48388)
This extension only implies decal support for OES textures. Remove it from the check for generic decal support.

In practice i think it would be unlikely that a driver supports this and not regular decal, but that would only make the bugs this may cause even harder to track down.