[Impeller] Refactor clip stack into separate testable class. (#51656)
The clip coverage tracking has had some bugs, but its been difficult to test as it was mixed into the regular entity pass workflow. This change pulls this logic and the clip recorder logic into a new class that is responsible for managing the coverage stacks.

Adds an unbalanced restore unit test as well.