Separate layer-specific data in DisplayListBuilder::SaveInfo (#53186)
Previously the SaveInfo structure contained data for both a regular Save and a SaveLayer call for each entry in the stack. This was not optimal in that it meant that data for the layers had to be copied back and forth on Save and Restore calls even if a layer was not involved.

Now Save stack entries share the data for their most recent enclosing layer and so updating that information does not require (error-prone and time-wasting) copying of the data between save stack entries. This should make new code that tracks new information for layers easier to add in the future.